// src/app/components/auth/model/user.model.ts
export interface UserModel {
  name: string;
  surename: string;
  userName: string;
  codiceFiscale: string;
  email: string;
  password: string;
  birthDay: string;
  role_user: boolean,
  role_admin: boolean,
  role_helpdesk: boolean,
  mobile: string,
  confirmed: boolean,
  verified: boolean
}